Consolidate MT-Bench runner ownership
Move the complete MT-Bench lifecycle into its registered runner and remove duplicate wrapper parameters.
